添加抽奖逻辑

+++ src/main/java/com/springboot/lottery/entity/PrizeEntity.java
@@ -18,6 +18,6 @@ public class PrizeEntity implements java.io.Serializable{
     private Integer surplusStock;
 
     //概率
-    private double percentage;
+    private Integer percentage;
 
 }
